refactor(core/notification/data-access): #232 use createFeature function for notification state feat
ISSUES CLOSED: #232
