+++ libs/core/notification/data-access/src/lib/+state/notification.facade.spec.ts
@@ -2,7 +2,7 @@ import { TestBed } from '@angular/core/testing';
 import { Store } from '@ngrx/store';
 import { provideMockStore } from '@ngrx/store/testing';
 
-import { dismissNotification } from './notification.actions';
+import { notificationActions } from './notification.actions';
 import { NotificationFacade } from './notification.facade';
 
 describe('NotificationFacade', () => {
@@ -25,7 +25,7 @@ describe('NotificationFacade', () => {
     });
 
     it('should dispatch dismissNotification action on dismiss', () => {
-      const action = dismissNotification();
+      const action = notificationActions.dismiss();
       facade.dismiss();
       expect(store.dispatch).toHaveBeenCalledWith(action);
     });
