+++ libs/core/notification/data-access/src/lib/+state/notification.facade.ts
@@ -1,15 +1,15 @@
 import { Injectable, inject } from '@angular/core';
 import { Store } from '@ngrx/store';
-import { dismissNotification } from './notification.actions';
-import { selectNotificationConfiguration } from './notification.selectors';
+import { notificationActions } from './notification.actions';
+import { selectConfiguration } from './notification.feature';
 
 @Injectable({ providedIn: 'root' })
 export class NotificationFacade {
   private readonly store = inject(Store);
 
-  config$ = this.store.select(selectNotificationConfiguration);
+  config$ = this.store.select(selectConfiguration);
 
   dismiss(): void {
-    this.store.dispatch(dismissNotification());
+    this.store.dispatch(notificationActions.dismiss());
   }
 }
