+++ libs/core/notification/data-access/src/lib/core-notification-data-access.module.ts
@@ -2,9 +2,9 @@ import { NgModule } from '@angular/core';
 import { EffectsModule } from '@ngrx/effects';
 import { StoreModule } from '@ngrx/store';
 import { NotificationEffects } from './+state/notification.effects';
-import { NOTIFICATION_FEATURE_KEY, notificationReducer } from './+state/notification.reducer';
+import { selectNotificationFeature } from './+state/notification.feature';
 
 @NgModule({
-  imports: [StoreModule.forFeature(NOTIFICATION_FEATURE_KEY, notificationReducer), EffectsModule.forFeature([NotificationEffects])],
+  imports: [StoreModule.forFeature(selectNotificationFeature), EffectsModule.forFeature([NotificationEffects])],
 })
 export class CoreNotificationDataAccessModule {}
